refactor: extract bp_brain to routes/brain.py
Phase 5.2 of the incremental modularisation. Pure mechanical move — zero
behaviour change, zero new endpoints.

## What moved

All 2 routes registered on `bp_brain`:

  /api/brain-history    — unified JSONL + log scan, returns list
  /api/brain-stream     — SSE tail of the same sources

## Pattern (same as bp_sessions)

- routes/brain.py declares `bp_brain = Blueprint('brain', __name__)` and
  the handlers, each starting with late `import dashboard as _d`.
- dashboard.py imports `bp_brain` from routes.brain and registers it
  alongside the other blueprints.
- Module-level helpers (`SESSIONS_DIR`, `SSE_MAX_SECONDS`, `_get_log_dirs`,
  `_tail_lines`, `_acquire_stream_slot`, `_release_stream_slot`,
  `_ext_emit`) stay in dashboard.py and are reached via `_d.`.

## Validation

Local live-fire (curl):
  /api/sessions        → 200, 716B
  /api/brain-history   → 200, 137KB  (total=307, 4 sources, 10 type buckets)
  /api/brain-stream    → 200 text/event-stream

Browser E2E (Chrome DevTools MCP): Overview + Brain render, events populate,
no regression.

## Sizes

  dashboard.py : 33,518 → 32,748 lines (-770)
  routes/brain.py: 799 lines (new)
